Moved all array transpose functions to LinAlg
Transposition is a concept of linear algebra rather than
multidimensional arrays of data.
