+++ base/linalg/bitarray.jl
@@ -211,3 +211,92 @@ function findmin(a::BitArray)
     k == l || return (false, ti)
     return m, mi
 end
+
+## transpose ##
+
+transpose(B::BitVector) = reshape(copy(B), 1, length(B))
+
+# fast 8x8 bit transpose from Henry S. Warrens's "Hacker's Delight"
+# http://www.hackersdelight.org/hdcodetxt/transpose8.c.txt
+function transpose8x8(x::UInt64)
+    y = x
+    t = xor(y, y >>> 7) & 0x00aa00aa00aa00aa
+    y = xor(y, t, t << 7)
+    t = xor(y, y >>> 14) & 0x0000cccc0000cccc
+    y = xor(y, t, t << 14)
+    t = xor(y, y >>> 28) & 0x00000000f0f0f0f0
+    return xor(y, t, t << 28)
+end
+
+function form_8x8_chunk(Bc::Vector{UInt64}, i1::Int, i2::Int, m::Int, cgap::Int, cinc::Int, nc::Int, msk8::UInt64)
+    x = UInt64(0)
+
+    k, l = Base.get_chunks_id(i1 + (i2 - 1) * m)
+    r = 0
+    for j = 1:8
+        k > nc && break
+        x |= ((Bc[k] >>> l) & msk8) << r
+        if l + 8 >= 64 && nc > k
+            r0 = 8 - Base._mod64(l + 8)
+            x |= (Bc[k + 1] & (msk8 >>> r0)) << (r + r0)
+        end
+        k += cgap + (l + cinc >= 64 ? 1 : 0)
+        l = Base._mod64(l + cinc)
+        r += 8
+    end
+    return x
+end
+
+# note: assumes B is filled with 0's
+function put_8x8_chunk(Bc::Vector{UInt64}, i1::Int, i2::Int, x::UInt64, m::Int, cgap::Int, cinc::Int, nc::Int, msk8::UInt64)
+    k, l = Base.get_chunks_id(i1 + (i2 - 1) * m)
+    r = 0
+    for j = 1:8
+        k > nc && break
+        Bc[k] |= ((x >>> r) & msk8) << l
+        if l + 8 >= 64 && nc > k
+            r0 = 8 - Base._mod64(l + 8)
+            Bc[k + 1] |= ((x >>> (r + r0)) & (msk8 >>> r0))
+        end
+        k += cgap + (l + cinc >= 64 ? 1 : 0)
+        l = Base._mod64(l + cinc)
+        r += 8
+    end
+    return
+end
+
+function transpose(B::BitMatrix)
+    l1 = size(B, 1)
+    l2 = size(B, 2)
+    Bt = falses(l2, l1)
+
+    cgap1, cinc1 = Base._div64(l1), Base._mod64(l1)
+    cgap2, cinc2 = Base._div64(l2), Base._mod64(l2)
+
+    Bc = B.chunks
+    Btc = Bt.chunks
+
+    nc = length(Bc)
+
+    for i = 1:8:l1
+        msk8_1 = UInt64(0xff)
+        if (l1 < i + 7)
+            msk8_1 >>>= i + 7 - l1
+        end
+
+        for j = 1:8:l2
+            x = form_8x8_chunk(Bc, i, j, l1, cgap1, cinc1, nc, msk8_1)
+            x = transpose8x8(x)
+
+            msk8_2 = UInt64(0xff)
+            if (l2 < j + 7)
+                msk8_2 >>>= j + 7 - l2
+            end
+
+            put_8x8_chunk(Btc, j, i, x, l2, cgap2, cinc2, nc, msk8_2)
+        end
+    end
+    return Bt
+end
+
+ctranspose(B::BitArray) = transpose(B)
